requirements.txt
numpy
matplotlib
Pillow
